HGF

Documentation for HGF.

HGF.calculate_auxiliary_prediction_precision

HGF.calculate_auxiliary_prediction_precision

HGF.calculate_posterior_mean

HGF.calculate_posterior_mean

HGF.calculate_posterior_mean_vape

HGF.calculate_posterior_mean_vape

HGF.calculate_posterior_mean_vope

HGF.calculate_posterior_precision

HGF.calculate_posterior_precision

HGF.calculate_posterior_precision_vape

HGF.calculate_posterior_precision_vape

HGF.calculate_posterior_precision_vope

HGF.calculate_prediction_mean

HGF.calculate_prediction_mean

HGF.calculate_prediction_precision

HGF.calculate_prediction_precision

HGF.calculate_prediction_precision

HGF.calculate_prediction_volatility

HGF.calculate_value_prediction_error

HGF.calculate_value_prediction_error

HGF.calculate_value_prediction_error

HGF.calculate_volatility_prediction_error

HGF.calculate_volatility_prediction_error

HGF.check_hgf

HGF.check_hgf

HGF.check_hgf

HGF.check_hgf

HGF.check_hgf

HGF.create_binary_softmax_response

HGF.create_gaussian_response

HGF.create_node

HGF.create_unit_square_sigmoid_response

HGF.enter_node_inputs!

HGF.enter_node_inputs!

HGF.enter_node_inputs!

HGF.give_inputs!

HGF.give_inputs!

HGF.give_inputs!

HGF.give_inputs!

HGF.give_inputs!

HGF.give_inputs!

HGF.init_hgf

HGF.premade_JGET

HGF.premade_agent

HGF.premade_binary_2level

HGF.premade_binary_3level

HGF.premade_continuous_2level

HGF.premade_hgf

HGF.update_hgf!

HGF.update_node_input!

HGF.update_node_posterior!

HGF.update_node_prediction!

HGF.update_node_prediction!

HGF.update_node_prediction!

HGF.update_node_prediction!

HGF.update_node_value_prediction_error!

HGF.update_node_value_prediction_error!

HGF.update_node_value_prediction_error!

HGF.update_node_volatility_prediction_error!

HGF.update_node_volatility_prediction_error!

HGF.update_node_volatility_prediction_error!

HGF.calculate_auxiliary_prediction_precision — Method

calculate_auxiliary_prediction_precision(self::AbstractInputNode)

An input nodes auxiliary precision is always 1.

HGF.calculate_auxiliary_prediction_precision — Method

calculate_auxiliary_prediction_precision(self::AbstractNode)

Calculates a node's auxiliary prediction precision.

HGF.calculate_posterior_mean — Method

calculate_posterior_mean(self::AbstractNode, value_children, volatility_children)

Calculates a node's posterior mean.

HGF.calculate_posterior_mean — Method

calculate_posterior_mean(self::BinaryStateNode, value_children, volatility_children)

Calculates a node's posterior mean.

HGF.calculate_posterior_mean_vape — Method

calculate_posterior_mean_vape(
    self::AbstractNode,
    child::AbstractNode)

Calculates the posterior mean update term for a single continuous value child to a state node.

HGF.calculate_posterior_mean_vape — Method

calculate_posterior_mean_vape(
    self::AbstractNode,
    child::BinaryStateNode)

Calculates the posterior mean update term for a single binary value child to a state node.

HGF.calculate_posterior_mean_vope — Method

calculate_posterior_mean_vope(
    self::AbstractNode,
    child::Any)

Calculates the posterior mean update term for a single continuos volatility child to a state node.

HGF.calculate_posterior_precision — Method

calculate_posterior_precision(
    self::AbstractNode,
    value_children,
    volatility_children)

Calculates a node's posterior precision.

HGF.calculate_posterior_precision — Method

calculate_posterior_precision(
    self::BinaryStateNode,
    value_children,
    volatility_children)

Calculates a binary node's posterior precision.

HGF.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(
    self::AbstractNode,
    child::AbstractNode)

Calculates the posterior precision update term for a single continuous value child to a state node.

HGF.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(
    self::AbstractNode,
    child::BinaryStateNode)

Calculates the posterior precision update term for a single binary value child to a state node.

HGF.calculate_posterior_precision_vope — Method

calculate_posterior_precision_vope(
    posterior_precision::Real,
    self::AbstractNode,
    volatility_children::Any)

Calculates the posterior precision update term for a single continuous volatility child to a state node.

HGF.calculate_prediction_mean — Method

calculate_prediction_mean(self::AbstractNode, value_parent::Any)

Calculates a node's prediction mean.

HGF.calculate_prediction_mean — Method

calculate_prediction_mean(self::AbstractNode, value_parent::Any)

Calculates a binary state node's prediction mean.

HGF.calculate_prediction_precision — Method

calculate_prediction_precision(self::InputNode)

Calculates an input node's prediction precision.

HGF.calculate_prediction_precision — Method

calculate_prediction_precision(self::StateNode)

Calculates a node's prediction precision.

HGF.calculate_prediction_precision — Method

calculate_prediction_precision(self::BinaryStateNode)

Calculates a binary state node's prediction precision.

HGF.calculate_prediction_volatility — Method

calculate_prediction_volatility(self::AbstractNode,  volatility_parents::Any)

Calculates a node's prediction volatility.

HGF.calculate_value_prediction_error — Method

calculate_value_prediction_error(self::AbstractNode)

Calculate's a state node's value prediction error.

HGF.calculate_value_prediction_error — Method

calculate_value_prediction_error(self::BinaryInputNode)

Calculates the prediciton error of a binary input node with finite precision

HGF.calculate_value_prediction_error — Method

calculate_value_prediction_error(self::InputNode, value_parents::Any)

Calculate's an input node's value prediction error.

HGF.calculate_volatility_prediction_error — Method

calculate_volatility_prediction_error(self::AbstractNode)

Calculates a state node's volatility prediction error.

HGF.calculate_volatility_prediction_error — Method

calculate_volatility_prediction_error(self::InputNode, value_parents::Any)

Calculates an input node's volatility prediction error.

HGF.check_hgf — Method

check_hgf(node::BinaryInputNode)

Function for checking the validity of a single binary input node

HGF.check_hgf — Method

check_hgf(node::BinaryStateNode)

Function for checking the validity of a single binary state node

HGF.check_hgf — Method

Function for checking if the specified HGF structure is valid

HGF.check_hgf — Method

check_hgf(node::InputNode)

Function for checking the validity of a single continuous input node

HGF.check_hgf — Method

check_hgf(node::StateNode)

Function for checking the validity of a single continous state node

HGF.create_binary_softmax_response — Method

create_binary_softmax_response(; state::String = "x1__prediction_mean")

Function for creating a customized binary softmax action model. Takes a state as argument. Outputs a function which inputs the state into a softmax, to get the action probability for a Bernoulli distribution.

HGF.create_gaussian_response — Method

create_gaussian_response(state::Vector{String})

Function for creating a customized gaussian response action model. Takes a node name and a string of states as arguments. Outputs a function which reports a linear combination of the chosen states from the chosen node with some noise.

HGF.create_node — Method

create_node(input_or_state_node, defaults, node_defaults, node_info)

Function for creating a node, given specifications

HGF.create_unit_square_sigmoid_response — Method

create_unit_square_sigmoid_response(; state::String = "x1__prediction_mean")

Function for creating a customized binary unit square sigmoid action model. Takes a state as argument. Outputs a function which inputs the state into a unit square sigmoid, to get the action probability for a Bernoulli distribution.

HGF.enter_node_inputs! — Method

enter_node_inputs!(hgf::HGFStruct, inputs::Dict)

Function for entering multiple inputs, structured as a dictionary, to multiple input nodes.

HGF.enter_node_inputs! — Method

enter_node_inputs!(hgf::HGFStruct, input::Number)

Function for entering a single input to a single input node. Can either take a single number, or a tuple which also includes the precision of the input.

HGF.enter_node_inputs! — Method

enter_node_inputs!(hgf::HGFStruct, inputs::Vector)

Function for entering multiple inputs, structured as a vector, to multiple input nodes.

HGF.give_inputs! — Method

give_inputs!(agent::AgentStruct, inputs::Array)

Function for inputting multiple observations to an agent. Input is structured as an array, with one column per input node and one row per timestep.

HGF.give_inputs! — Method

give_inputs!(agent::AgentStruct, inputs::Dict{String,Vector})

Function for inputting multiple observations to an agent. Input is structured as a dictionary with a vector for each input node.

HGF.give_inputs! — Method

give_inputs!(agent::AgentStruct, inputs::Number)

Function for giving multiple inputs to an AgentStruct. Here a single number is passed as an input.

HGF.give_inputs! — Method

give_inputs!(hgf::HGFStruct, inputs::Array)

Function for inputting multiple observations to an hgf. Input is structured as an array, with one column per input node and one row per timestep.

HGF.give_inputs! — Method

give_inputs!(hgf::HGFStruct, inputs::Dict{String,Vector})

Function for inputting multiple observations to an hgf. Input is structured as a dictionary with a vector for each input node.

HGF.give_inputs! — Method

give_inputs!(hgf::HGFStruct, inputs::Number)

Function for inputting multiple observations to an hgf. Input is a single value.

HGF.init_hgf — Method

function init_hgf(
    default_params,
    input_nodes,
    state_nodes,
    edges,
    update_order = false,
)

Function for initializing the structure of an HGF model.

HGF.premade_JGET — Function

premade_JGET(params_list, starting_state_list)

The JGET model. It has a single continuous input node u, with a value parent x1, and a volatility parent x3. x1 has volatility parent x2, and x3 has a volatility parent x4.

HGF.premade_agent — Function

function premade_agent(
    model_name::String,
    perception_model = (;),
    params = Dict(),
    states = Dict(),
    specifications = (;),
)

Function for initializing the structure of an agent model.

HGF.premade_binary_2level — Method

premade_binary_2level(params_list, starting_state_list)

The standard binary 2 level HGF model

HGF.premade_binary_3level — Method

premade_binary_3level(params_list, starting_state_list)

The standard binary 3 level HGF model

HGF.premade_continuous_2level — Method

premade_continuous_2level(params_list, starting_state_list)

The standard 2 level HGF. It has a continous input node U, with a single value parent x1, which in turn has a single volatility parent x2.

HGF.premade_hgf — Function

function premade_hgf(
    model_name,
    params,
    starting_state,
)

Function for initializing the structure of an HGF model.

HGF.update_hgf! — Method

update_hgf!(hgf::HGFStruct, inputs)

Function for updating all nodes in an HGF hierarchy.

HGF.update_node_input! — Method

update_node_input!(self::InputNode, input::Real)

Function for updating the input for a single input node

HGF.update_node_posterior! — Method

update_node_posterior!(self::StateNode)

Function for updating the posterior of a single state node

HGF.update_node_prediction! — Method

update_node_prediction!(self::InputNode)

Function for updating the posterior of a single input node

HGF.update_node_prediction! — Method

update_node_prediction!(self::StateNode)

Function for updating the prediction for a single state node

HGF.update_node_prediction! — Method

update_node_prediction!(self::BinaryInputNode)

The prediction precision is constant for binary input nodes, so nothing is done here.

HGF.update_node_prediction! — Method

update_node_prediction!(self::BinaryStateNode)

Function for updating the prediction for a single Binary state node

HGF.update_node_value_prediction_error! — Method

update_node_value_prediction_error!(self::InputNode)

Function for updating the value prediction error of a single input node

HGF.update_node_value_prediction_error! — Method

Function for updating the value prediction error of a single state node

HGF.update_node_value_prediction_error! — Method

update_node_value_prediction_error!(self::BinaryInputNode)

Function for updating the value prediction error of a single input node.

HGF.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(self::InputNode)

Function for updating the volatility prediction error of a single input node

HGF.update_node_volatility_prediction_error! — Method

Function for updating the volatility prediction error of a single state node

HGF.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(self::BinaryInputNode)

There is no volatility prediction error for binary input nodes.