Overview of all functions in the Hierarchical Gaussian Filtering package

HierarchicalGaussianFiltering.BinaryInputNode — Type

HierarchicalGaussianFiltering.BinaryInputNodeHistory — Type

Configuration of history of binary input node

HierarchicalGaussianFiltering.BinaryInputNodeParameters — Type

Configuration of parameters in binary input node. Default category mean set to [0,1]

HierarchicalGaussianFiltering.BinaryInputNodeState — Type

Configuration of states of binary input node

HierarchicalGaussianFiltering.BinaryStateNode — Type

Overview of edge posibilities

HierarchicalGaussianFiltering.BinaryStateNodeHistory — Type

Overview of the history of the binary state node

HierarchicalGaussianFiltering.BinaryStateNodeParameters — Type

Configure parameters of binary state node

HierarchicalGaussianFiltering.BinaryStateNodeState — Type

Overview of the states of the binary state node

HierarchicalGaussianFiltering.CategoricalInputNode — Type

HierarchicalGaussianFiltering.CategoricalInputNodeHistory — Type

History of categorical input node

HierarchicalGaussianFiltering.CategoricalInputNodeState — Type

Configuration of states of categorical input node

HierarchicalGaussianFiltering.CategoricalStateNode — Type

Configuration of edges in categorical state node

HierarchicalGaussianFiltering.CategoricalStateNodeHistory — Type

Configuration of history in categorical state node

HierarchicalGaussianFiltering.CategoricalStateNodeState — Type

Configuration of states in categorical state node

HierarchicalGaussianFiltering.ContinuousInputNode — Type

HierarchicalGaussianFiltering.ContinuousInputNodeHistory — Type

Configuration of continuous input node history

HierarchicalGaussianFiltering.ContinuousInputNodeParameters — Type

Configuration of continuous input node parameters

HierarchicalGaussianFiltering.ContinuousInputNodeState — Type

Configuration of continuous input node states

HierarchicalGaussianFiltering.ContinuousStateNode — Type

HierarchicalGaussianFiltering.ContinuousStateNodeHistory — Type

Configuration of continuous state node history

HierarchicalGaussianFiltering.ContinuousStateNodeParameters — Type

Configuration of continuous state nodes' parameters

HierarchicalGaussianFiltering.ContinuousStateNodeState — Type

Configurations of the continuous state node states

HierarchicalGaussianFiltering.HGF — Type

HierarchicalGaussianFiltering.OrderedNodes — Type

ActionModels.get_history — Method

get_history(hgf::HGF, target_state::Tuple{String,String})

Gets the history of a state from a specific node in an HGF. A vector of states can also be passed.

get_history(hgf::HGF, node_name::String)

Gets the history of all states for a specific node in an HGF. If only a node object is passed, it will return the history of all states in that node. If only an HGF object is passed, it will return the history of all states in all nodes in the HGF.

ActionModels.get_parameters — Method

get_parameters(hgf::HGF, target_param::Tuple{String,String})

Gets a single parameter value from a specific node in an HGF. A vector of parameters can also be passed.

get_parameters(hgf::HGF, node_name::String)

Gets all parameter values for a specific node in an HGF. If only a node object is passed, returns all parameters in that node. If only an HGF object is passed, returns all parameters of all nodes in the HGF.

ActionModels.get_states — Method

get_states(hgf::HGF, target_state::Tuple{String,String})

Gets a single state value from a specific node in an HGF. A vector of states can also be passed.

get_states(hgf::HGF, node_name::String)

Gets all parameter values for a specific node in an HGF. If only a node object is passed, returns all states in that node. If only an HGF object is passed, returns all states of all nodes in the HGF.

ActionModels.give_inputs! — Method

give_inputs!(hgf::HGF, inputs)

Give inputs to an agent. Input can be a single value, a vector of values, or an array of values.

ActionModels.plot_predictive_simulation — Method

plot_predictive_simulation(hgf::HGF, parameter_distributions, target_state, inputs; kwargs...)

Runs and plots results from a predictive simulation using only an HGF, instead of an agent. See the ActionModels documentation for more information.

ActionModels.plot_trajectory — Method

plot_trajectory(hgf::HGF, node_name::String; kwargs...)

Plots the trajectory of a node in an HGF. See the ActionModels documentation for more information.

ActionModels.premade_agent — Function

premade_agent(model_name::String, hgf::HGF, config::Dict = Dict(); verbose = true)

Create an agent fom the list of premade agents. If an HGF is passed as a separate argument, add it to the config dictionary.

ActionModels.reset! — Method

reset!(hgf::HGF)

Reset an HGF to its initial state.

ActionModels.set_parameters! — Method

set_parameters!(hgf::HGF, target_param::Tuple, param_value::Any)

Setting a single parameter value for an HGF.

set_parameters!(hgf::HGF, parameter_values::Dict)

Set mutliple parameters values for an HGF. Takes a dictionary of parameter names and values.

HierarchicalGaussianFiltering.calculate_auxiliary_prediction_precision — Method

calculate_auxiliary_prediction_precision(node::AbstractInputNode)

An input node's auxiliary prediction precision is always 1.

HierarchicalGaussianFiltering.calculate_auxiliary_prediction_precision — Method

calculate_auxiliary_prediction_precision(node::AbstractNode)

Calculates a node's auxiliary prediction precision.

Uses the equation $\gamma_i = \nu_i \cdot \hat{\pi}_i$

HierarchicalGaussianFiltering.calculate_posterior — Method

calculate_posterior(node::CategoricalStateNode)

Calculate the posterior for a categorical state node.

One hot encoding $\vec{u} = [0, 0, \dots ,1, \dots,0]$

HierarchicalGaussianFiltering.calculate_posterior_mean — Method

calculate_posterior_mean(node::AbstractNode)

Calculates a node's posterior mean.

Uses the equation $$

HierarchicalGaussianFiltering.calculate_posterior_mean — Method

calculate_posterior_mean(node::BinaryStateNode)

Calculates a node's posterior mean.

Uses the equation $\mu = \frac{e^{-0.5 \cdot \pi_n \cdot \eta_1^2}}{\hat{\mu}_n \cdot e^{-0.5 \cdot \pi_n \cdot \eta_1^2} \; + 1-\hat{\mu}_n \cdot e^{-0.5 \cdot \pi_n \cdot \eta_2^2}}$

HierarchicalGaussianFiltering.calculate_posterior_mean_value_child_increment — Method

calculate_posterior_mean_value_child_increment(node::AbstractNode, child::AbstractNode)

Calculates the posterior mean update term for a single continuous value child to a state node.

Uses the equation $$

HierarchicalGaussianFiltering.calculate_posterior_mean_value_child_increment — Method

calculate_posterior_mean_value_child_increment(node::AbstractNode, child::BinaryStateNode)

Calculates the posterior mean update term for a single binary value child to a state node.

Uses the equation $$

HierarchicalGaussianFiltering.calculate_posterior_mean_volatility_child_increment — Method

calculate_posterior_mean_volatility_child_increment(node::AbstractNode, child::AbstractNode)

Calculates the posterior mean update term for a single continuos volatility child to a state node.

Uses the equation $$

HierarchicalGaussianFiltering.calculate_posterior_precision — Method

calculate_posterior_precision(node::AbstractNode)

Calculates a node's posterior precision.

Uses the equation $\pi_i^{'} = \hat{\pi}_i +\underbrace{\sum_{j=1}^{j\;children} \alpha_{j,i} \cdot \hat{\pi}_{j}} _\text{sum \;of \;VAPE \; continuous \; value \;chidren}$

HierarchicalGaussianFiltering.calculate_posterior_precision — Method

calculate_posterior_precision(node::BinaryStateNode)

Calculates a binary node's posterior precision.

Uses the equations

$\pi_n = inf$ if the precision is infinite

$\pi_n = \frac{1}{\hat{\mu}_n \cdot (1-\hat{\mu}_n)}$ if the precision is other than infinite

HierarchicalGaussianFiltering.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(node::AbstractNode, child::AbstractNode)

Calculates the posterior precision update term for a single continuous value child to a state node.

Uses the equation $$

HierarchicalGaussianFiltering.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(node::AbstractNode, child::BinaryStateNode)

Calculates the posterior precision update term for a single binary value child to a state node.

Uses the equation $$

HierarchicalGaussianFiltering.calculate_posterior_precision_vope — Method

calculate_posterior_precision_vope(node::AbstractNode, child::AbstractNode)

Calculates the posterior precision update term for a single continuous volatility child to a state node.

Uses the equation $$

HierarchicalGaussianFiltering.calculate_prediction — Method

function calculate_prediction(node::CategoricalStateNode)

Calculate the prediction for a categorical state node.

Uses the equation $\vec{\hat{\mu}_n}= \frac{\hat{\mu}_j}{\sum_{j=1}^{j\;binary \;parents} \hat{\mu}_j}$

HierarchicalGaussianFiltering.calculate_prediction_mean — Method

calculate_prediction_mean(node::AbstractNode)

Calculates a node's prediction mean.

Uses the equation $\hat{\mu}_i=\mu_i+\sum_{j=1}^{j\;value\;parents} \mu_{j} \cdot \alpha_{i,j}$

HierarchicalGaussianFiltering.calculate_prediction_mean — Method

calculate_prediction_mean(node::BinaryStateNode)

Calculates a binary state node's prediction mean.

Uses the equation $\hat{\mu}_n= \big(1+e^{\sum_{j=1}^{j\;value \; parents} \hat{\mu}_{j}}\big)^{-1}$

HierarchicalGaussianFiltering.calculate_prediction_precision — Method

calculate_prediction_precision(node::AbstractInputNode)

Calculates an input node's prediction precision.

Uses the equation $\hat{\pi}_n = \frac{1}{\nu}_n$

HierarchicalGaussianFiltering.calculate_prediction_precision — Method

calculate_prediction_precision(node::AbstractNode)

Calculates a node's prediction precision.

Uses the equation $\hat{\pi}_i^ = \frac{1}{\frac{1}{\pi_i}+\nu_i^}$

HierarchicalGaussianFiltering.calculate_prediction_precision — Method

calculate_prediction_precision(node::BinaryStateNode)

Calculates a binary state node's prediction precision.

Uses the equation $\hat{\pi}_n = \frac{1}{\hat{\mu}_n \cdot (1-\hat{\mu}_n)}$

HierarchicalGaussianFiltering.calculate_prediction_volatility — Method

calculate_prediction_volatility(node::AbstractNode)

Calculates a node's prediction volatility.

Uses the equation $\nu_i =exp( \omega_i + \sum_{j=1}^{j\;volatility\;parents} \mu_{j} \cdot \kappa_{i,j}}$

HierarchicalGaussianFiltering.calculate_value_prediction_error — Method

calculate_value_prediction_error(node::AbstractNode)

Calculate's a state node's value prediction error.

Uses the equation $\delta_n = \mu_n - \hat{\mu}_n$

HierarchicalGaussianFiltering.calculate_value_prediction_error — Method

calculate_value_prediction_error(node::BinaryInputNode)

Calculates the prediciton error of a binary input node with finite precision.

Uses the equation $\delta_n= u - \sum_{j=1}^{j\;value\;parents} \hat{\mu}_{j}$

HierarchicalGaussianFiltering.calculate_value_prediction_error — Method

calculate_value_prediction_error(node::CategoricalStateNode)

Calculate the value prediction error for a categorical state node.

Uses the equation $\delta_n= u - \sum_{j=1}^{j\;value\;parents} \hat{\mu}_{j}$

HierarchicalGaussianFiltering.calculate_value_prediction_error — Method

calculate_value_prediction_error(node::ContinuousInputNode)

Calculate's an input node's value prediction error.

Uses the equation $\delta_n= u - \sum_{j=1}^{j\;value\;parents} \hat{\mu}_{j}$

HierarchicalGaussianFiltering.calculate_volatility_prediction_error — Method

calculate_volatility_prediction_error(node::AbstractNode)

Calculates a state node's volatility prediction error.

Uses the equation $\Delta_n = \frac{\hat{\pi}_n}{\pi_n} + \hat{\pi}_n \cdot \delta_n^2-1$

HierarchicalGaussianFiltering.calculate_volatility_prediction_error — Method

calculate_volatility_prediction_error(node::ContinuousInputNode)

Calculates an input node's volatility prediction error.

Uses the equation $\mu'_j=\sum_{j=1}^{j\;value\;parents} \mu_{j}$ $\pi'_j=\frac{{\sum_{j=1}^{j\;value\;parents} \pi_{j}}}{j}$ $\Delta_n=\frac{\hat{\pi}_n}{\pi'_j} + \hat{\mu}_i\cdot (u -\mu'_j^2 )-1$

HierarchicalGaussianFiltering.check_hgf — Method

check_hgf(hgf::HGF)

Check whether an HGF has specified correctly. A single node can also be passed.

HierarchicalGaussianFiltering.enter_node_inputs! — Method

enter_node_inputs!(hgf::HGF, input)

Set input values in input nodes. Can either take a single value, a vector of values, or a dictionary of input node names and corresponding values.

HierarchicalGaussianFiltering.get_prediction — Function

get_prediction(hgf::HGF, node_name::String)

Get the prediction for the next timestep for a specified node in an HGF. If an agent is passed instead of an HGF, the HGF is extracted from the substruct in the agent. A single node can also be passed.

HierarchicalGaussianFiltering.get_surprise — Function

getsurprise(hgf::HGF, nodename::String = "u")

Calculates the surprisal at the last input of a specified input node in an HGF. If an agent is passed instead of an HGF, the HGF is extracted from the substruct in the agent. If no node is specified, the surprisal of all input nodes is summed.

HierarchicalGaussianFiltering.get_surprise — Method

get_surprise(node::CategoricalInputNode)

Calculate the surprise of a categorical input node on seeing the last input.

HierarchicalGaussianFiltering.get_surprise — Method

get_surprise(node::ContinuousInputNode)

Calculate the surprise of an input node on seeing the last input.

Equation: $\hat{\mu}'_j={\sum_{j=1}^{j\;value\;parents} \hat{\mu}_j}$

$\Im= -log(pdf(\mathcal{N}(\hat{\mu}'_j, \hat{\pi}_j), u))$

HierarchicalGaussianFiltering.hgf_binary_softmax_action — Method

hgf_binary_softmax_action(agent, input)

Action model which gives a binary action. The action probability is the softmax of a specified state of a node.

In addition to the HGF substruct, the following must be present in the agent: Parameters: "softmaxactionprecision" Settings: "target_state"

HierarchicalGaussianFiltering.hgf_gaussian_action — Method

hgf_gaussian_action(agent::Agent, input)

Action model which reports a given HGF state with Gaussian noise.

In addition to the HGF substruct, the following must be present in the agent: Parameters: "gaussianactionprecision" Settings: "target_state"

HierarchicalGaussianFiltering.hgf_predict_category_action — Method

hgf_predict_category_action(agent::Agent, input)

Action model which gives a categorical prediction of the input, based on an HGF. The HGF used must be a categorical HGF.

In addition to the HGF substruct, the following must be present in the agent: Settings: "targetcategoricalnode"

HierarchicalGaussianFiltering.hgf_unit_square_sigmoid_action — Method

unit_square_sigmoid_action(agent, input)

Action model which gives a binary action. The action probability is the unit square sigmoid of a specified state of a node.

In addition to the HGF substruct, the following must be present in the agent: Parameters: "sigmoidactionprecision" Settings: "target_state"

HierarchicalGaussianFiltering.init_hgf — Method

init_hgf(;
    input_nodes::Union{String,Dict,Vector},
    state_nodes::Union{String,Dict,Vector},
    edges::Union{Vector{<:Dict},Dict},
    node_defaults::Dict = Dict(),
    update_order::Union{Nothing,Vector{String}} = nothing,
    verbose::Bool = true,
)

Initialize an HGF. Node information includes 'name' and 'type' as keys, as well as any other parameters that are specific to the node type. Edge information includes 'child', as well as 'valueparents' and/or 'volatilityparents' as keys. Parents are vectors of either node name strings, or tuples with node names and coupling strengths.

Arguments

'input_nodes::Union{String,Dict,Vector}': Input nodes to be created. Can either be a string with a node name, a dictionary with node information, or a vector of strings and/or dictionaries.

'state_nodes::Union{String,Dict,Vector}': State nodes to be created. Can either be a string with a node name, a dictionary with node information, or a vector of strings and/or dictionaries.

'edges::Union{Vector{<:Dict},Dict}': Edges to be created. Can either be a dictionary with edge information, or a vector of dictionaries.

'node_defaults::Dict = Dict()': A dictionary with default values for the nodes. If a node is created without specifying a value for a parameter, the default value is used.

'update_order::Union{Nothing,Vector{String}} = nothing': The order in which the nodes are updated. If set to nothing, the update order is determined automatically.

'verbose::Bool = true': If set to false, warnings are hidden.

Examples

##Create a simple 2level continuous HGF##

#List of input nodes
input_nodes = Dict(
    "name" => "u",
    "type" => "continuous",
    "evolution_rate" => -2,
)

#List of state nodes
state_nodes = [
    Dict(
        "name" => "x1",
        "type" => "continuous",
        "evolution_rate" => -2,
        "initial_mean" => 0,
        "initial_precision" => 1,
    ),
    Dict(
        "name" => "x2",
        "type" => "continuous",
        "evolution_rate" => -2,
        "initial_mean" => 0,
        "initial_precision" => 1,
    ),
]

#List of child-parent relations
edges = [
    Dict(
        "child" => "u",
        "value_parents" => ("x1", 1),
    ),
    Dict(
        "child" => "x1",
        "volatility_parents" => ("x2", 1),
    ),
]

#Initialize the HGF
hgf = init_hgf(
    input_nodes = input_nodes,
    state_nodes = state_nodes,
    edges = edges,
)

##Create a more complicated HGF without specifying information for each node##

#Set defaults for all nodes
node_defaults = Dict(
    "evolution_rate" => -2,
    "initial_mean" => 0,
    "initial_precision" => 1,
    "value_coupling" => 1,
    "volatility_coupling" => 1,
)

input_nodes = [
    "u1",
    "u2",
]

state_nodes = [
    "x1",
    "x2",
    "x3",
    "x4",
]

edges = [
    Dict(
        "child" => "u1",
        "value_parents" => ["x1", "x2"],
        "volatility_parents" => "x3"
    ),
    Dict(
        "child" => "u2",
        "value_parents" => ["x1"],
    ),
    Dict(
        "child" => "x1",
        "volatility_parents" => "x4",
    ),
    Dict(
        "child" => "x2",
        "volatility_parents" => "x4",
    ),
]

hgf = init_hgf(
    input_nodes = input_nodes,
    state_nodes = state_nodes,
    edges = edges,
    node_defaults = node_defaults,
)

HierarchicalGaussianFiltering.init_node — Method

init_node(input_or_state_node, node_defaults, node_info)

Function for creating a node, given specifications

HierarchicalGaussianFiltering.premade_JGET — Method

premade_JGET(config::Dict; verbose::Bool = true)

The HGF used in the JGET model. It has a single continuous input node u, with a value parent x1, and a volatility parent x3. x1 has volatility parent x2, and x3 has a volatility parent x4.

Config defaults:

("u", "evolution_rate"): -2

("x1", "evolution_rate"): -2

("x2", "evolution_rate"): -2

("x3", "evolution_rate"): -2

("x4", "evolution_rate"): -2

("u", "x1", "value_coupling"): 1

("u", "x3", "value_coupling"): 1

("x1", "x2", "volatility_coupling"): 1

("x3", "x4", "volatility_coupling"): 1

("x1", "initial_mean"): 0

("x1", "initial_precision"): 1

("x2", "initial_mean"): 0

("x2", "initial_precision"): 1

("x3", "initial_mean"): 0

("x3", "initial_precision"): 1

("x4", "initial_mean"): 0

("x4", "initial_precision"): 1

HierarchicalGaussianFiltering.premade_binary_2level — Method

premade_binary_2level(config::Dict; verbose::Bool = true)

The standard binary 2 level HGF model, which takes a binary input, and learns the probability of either outcome. It has one binary input node u, with a binary value parent x1, which in turn has a continuous value parent x2.

Config defaults:

("u", "category_means"): [0, 1]

("u", "input_precision"): Inf

("x2", "evolution_rate"): -2

("x1", "x2", "value_coupling"): 1

("x2", "initial_mean"): 0

("x2", "initial_precision"): 1

HierarchicalGaussianFiltering.premade_binary_3level — Method

premade_binary_3level(config::Dict; verbose::Bool = true)

The standard binary 3 level HGF model, which takes a binary input, and learns the probability of either outcome. It has one binary input node u, with a binary value parent x1, which in turn has a continuous value parent x2. This then has a continunous volatility parent x3.

This HGF has five shared parameters: "x2evolutionrates" "x2initialprecisions" "x2initialmeans" "valuecouplingsx1x2" "volatilitycouplingsx2x3"

Config defaults:

("u", "category_means"): [0, 1]

("u", "input_precision"): Inf

("x2", "evolution_rate"): -2

("x3", "evolution_rate"): -2

("x1", "x2", "value_coupling"): 1

("x2", "x3", "volatility_coupling"): 1

("x2", "initial_mean"): 0

("x2", "initial_precision"): 1

("x3", "initial_mean"): 0

("x3", "initial_precision"): 1

HierarchicalGaussianFiltering.premade_categorical_3level — Method

premade_categorical_3level(config::Dict; verbose::Bool = true)

The categorical 3 level HGF model, which takes an input from one of n categories and learns the probability of a category appearing. It has one categorical input node u, with a categorical value parent x1. The categorical node has a binary value parent x1n for each category n, each of which has a continuous value parent x2n. Finally, all of these continuous nodes share a continuous volatility parent x3. Setting parameter values for x1 and x2 sets that parameter value for each of the x1n and x2n nodes.

Config defaults:

"n_categories": 4

("x2", "evolution_rate"): -2

("x3", "evolution_rate"): -2

("x1", "x2", "value_coupling"): 1

("x2", "x3", "volatility_coupling"): 1

("x2", "initial_mean"): 0

("x2", "initial_precision"): 1

("x3", "initial_mean"): 0

("x3", "initial_precision"): 1

HierarchicalGaussianFiltering.premade_categorical_3level_state_transitions — Method

premade_categorical_3level_state_transitions(config::Dict; verbose::Bool = true)

The categorical state transition 3 level HGF model, learns state transition probabilities between a set of n categorical states. It has one categorical input node u, with a categorical value parent x1n for each of the n categories, representing which category was transitioned from. Each categorical node then has a binary parent x1nm, representing the category m which the transition was towards. Each binary node x1nm has a continuous parent x2nm. Finally, all of these continuous nodes share a continuous volatility parent x3. Setting parameter values for x1 and x2 sets that parameter value for each of the x1nm and x2n_m nodes.

This HGF has five shared parameters: "x2evolutionrates" "x2initialprecisions" "x2initialmeans" "valuecouplingsx1x2" "volatilitycouplingsx2x3"

Config defaults:

- "n_categories": 4
- ("x2", "evolution_rate"): -2
- ("x3", "evolution_rate"): -2
- ("x1", "x2", "volatility_coupling"): 1
- ("x2", "x3", "volatility_coupling"): 1
- ("x2", "initial_mean"): 0
- ("x2", "initial_precision"): 1
- ("x3", "initial_mean"): 0
- ("x3", "initial_precision"): 1

HierarchicalGaussianFiltering.premade_continuous_2level — Method

premade_continuous_2level(config::Dict; verbose::Bool = true)

The standard 2 level continuous HGF, which filters a continuous input. It has a continous input node u, with a single value parent x1, which in turn has a single volatility parent x2.

Config defaults:

("u", "evolution_rate"): -2

("x1", "evolution_rate"): -2

("x2", "evolution_rate"): -2

("u", "x1", "value_coupling"): 1

("x1", "x2", "volatility_coupling"): 1

("x1", "initial_mean"): 0

("x1", "initial_precision"): 1

("x2", "initial_mean"): 0

("x2", "initial_precision"): 1

HierarchicalGaussianFiltering.premade_hgf — Function

premade_hgf(model_name::String, config::Dict = Dict(); verbose = true)

Create an HGF from the list of premade HGFs.

Arguments

'model_name::String': Name of the premade HGF. Returns a list of possible model names if set to 'help'.

'config::Dict = Dict()': A dictionary with configurations for the HGF, like parameters and settings.

'verbose::Bool = true': If set to false, warnings are hidden.

HierarchicalGaussianFiltering.premade_hgf_binary_softmax — Method

premade_hgf_binary_softmax(config::Dict)

Create an agent suitable for the HGF binary softmax model.

Config defaults:

"HGF": "binary_3level"

"softmaxactionprecision": 1

"targetstate": ("x1", "predictionmean")

HierarchicalGaussianFiltering.premade_hgf_gaussian — Method

premade_hgf_gaussian(config::Dict)

Create an agent suitable for the HGF Gaussian action model.

Config defaults:

"HGF": "continuous_2level"

"gaussianactionprecision": 1

"targetstate": ("x1", "posteriormean")

HierarchicalGaussianFiltering.premade_hgf_multiple_actions — Method

premade_hgf_multiple_actions(config::Dict)

Create an agent suitable for multiple aciton models that can depend on an HGF substruct. The used action models are specified as a vector.

Config defaults:

"HGF": "continuous_2level"

"hgfactions": ["gaussianaction", "softmaxaction", "unitsquaresigmoidaction"]

HierarchicalGaussianFiltering.premade_hgf_predict_category — Method

premade_hgf_predict_category(config::Dict)

Create an agent suitable for the HGF predict category model.

Config defaults:

"HGF": "categorical_3level"

"targetcategoricalnode": "x1"

HierarchicalGaussianFiltering.premade_hgf_unit_square_sigmoid — Method

premade_hgf_binary_softmax(config::Dict)

Create an agent suitable for the HGF unit square sigmoid model.

Config defaults:

"HGF": "binary_3level"

"sigmoidactionprecision": 1

"targetstate": ("x1", "predictionmean")

HierarchicalGaussianFiltering.update_hgf! — Method

update_hgf!(
    hgf::HGF,
    inputs::Union{
        Real,
        Missing,
        Vector{<:Union{Real,Missing}},
        Dict{String,<:Union{Real,Missing}},
    },
)

Update all nodes in an HGF based on an input. The input can either be missing, a single value, a vector of values, or a dictionary of input node names and corresponding values.

HierarchicalGaussianFiltering.update_hgf_binary_softmax_action — Method

update_hgf_softmax_action(agent::Agent, input)

Action model that first updates the HGF, and then passes a state from the HGF through a softmax to give a binary action.

In addition to the HGF substruct, the following must be present in the agent: Parameters: "softmaxactionprecision" Settings: "target_state"

HierarchicalGaussianFiltering.update_hgf_gaussian_action — Method

update_hgf_gaussian_action(agent::Agent, input)

Action model that first updates the HGF, and then reports a given HGF state with Gaussian noise.

In addition to the HGF substruct, the following must be present in the agent: Parameters: "gaussianactionprecision" Settings: "target_state"

HierarchicalGaussianFiltering.update_hgf_multiple_actions — Method

update_hgf_multiple_actions(agent::Agent, input)

Action model that first updates the HGF, and then runs multiple action models.

HierarchicalGaussianFiltering.update_hgf_predict_category_action — Method

update_hgf_predict_category_action(agent::Agent, input)

Action model that first updates the HGF, and then returns a categorical prediction of the input. The HGF used must be a categorical HGF.

In addition to the HGF substruct, the following must be present in the agent: Settings: "targetcategoricalnode"

HierarchicalGaussianFiltering.update_hgf_unit_square_sigmoid_action — Method

update_hgf_unit_square_sigmoid_action(agent::Agent, input)

Action model that first updates the HGF, and then passes a state from the HGF through a unit square sigmoid transform to give a binary action.

In addition to the HGF substruct, the following must be present in the agent: Parameters: "sigmoidactionprecision" Settings: "target_state"

HierarchicalGaussianFiltering.update_node_input! — Method

update_node_input!(node::AbstractInputNode, input::Union{Real,Missing})

Update the prediction of a single input node.

HierarchicalGaussianFiltering.update_node_posterior! — Method

update_node_posterior!(node::AbstractStateNode)

Update the posterior of a single continuous state node.

HierarchicalGaussianFiltering.update_node_posterior! — Method

update_node_posterior!(node::CategoricalStateNode)

Update the posterior of a single binary state node.

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::AbstractInputNode)

Update the posterior of a single input node.

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::AbstractStateNode)

Update the prediction of a single state node.

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::BinaryInputNode)

There is no prediction update for binary input nodes, as the prediction precision is constant.

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::BinaryStateNode)

Update the prediction of a single binary state node.

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::CategoricalInputNode)

There is no prediction update for categorical input nodes, as the prediction precision is constant.

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::CategoricalStateNode)

Update the prediction of a single categorical state node.

HierarchicalGaussianFiltering.update_node_value_prediction_error! — Method

update_node_value_prediction_error!(node::AbstractInputNode)

Update the value prediction error of a single input node.

HierarchicalGaussianFiltering.update_node_value_prediction_error! — Method

update_node_value_prediction_error!(node::AbstractStateNode)

Update the value prediction error of a single state node.

HierarchicalGaussianFiltering.update_node_value_prediction_error! — Method

update_node_value_prediction_error!(node::BinaryInputNode)

Update the value prediction error of a single binary input node.

HierarchicalGaussianFiltering.update_node_value_prediction_error! — Method

update_node_value_prediction_error!(node::CategoricalInputNode)

There is no value prediction error update for categorical input nodes.

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(node::AbstractInputNode)

Update the value prediction error of a single input node.

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(node::AbstractStateNode)

Update the volatility prediction error of a single state node.

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(node::BinaryInputNode)

There is no volatility prediction error update for binary input nodes.

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(node::BinaryStateNode)

There is no volatility prediction error update for binary state nodes.

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(node::CategoricalInputNode)

There is no volatility prediction error update for categorical input nodes.

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(node::CategoricalStateNode)

There is no volatility prediction error update for categorical state nodes.

This page was generated using Literate.jl.