HGF

Documentation for HGF.

HGF.calculate_auxiliary_prediction_precision

HGF.calculate_posterior_mean

HGF.calculate_posterior_mean_vape

HGF.calculate_posterior_mean_vape

HGF.calculate_posterior_mean_vape

HGF.calculate_posterior_mean_vope

HGF.calculate_posterior_mean_vope

HGF.calculate_posterior_mean_vope

HGF.calculate_posterior_precision

HGF.calculate_posterior_precision_vape

HGF.calculate_posterior_precision_vape

HGF.calculate_posterior_precision_vape

HGF.calculate_posterior_precision_vope

HGF.calculate_posterior_precision_vope

HGF.calculate_posterior_precision_vope

HGF.calculate_posterior_precision_vope_helper

HGF.calculate_prediction_mean

HGF.calculate_prediction_mean

HGF.calculate_prediction_mean

HGF.calculate_prediction_precision

HGF.calculate_prediction_volatility

HGF.calculate_prediction_volatility

HGF.calculate_prediction_volatility

HGF.calculate_value_prediction_error

HGF.calculate_volatility_prediction_error

HGF.update_node

HGF.update_node

HGF.calculate_auxiliary_prediction_precision — Method

calculate_auxiliary_prediction_precision(self::Node)

Calculates a node's auxiliary prediction precision.

HGF.calculate_posterior_mean — Method

calculate_posterior_mean(self::Node, value_children, volatility_children)

Calculates a node's posterior mean.

HGF.calculate_posterior_mean_vape — Method

calculate_posterior_mean_vape(posterior_mean::AbstractFloat, self::Node, value_child::Bool)

Calculates a node's posterior mean for a VAPE coupling without a value child.

HGF.calculate_posterior_mean_vape — Method

calculate_posterior_mean_vape(posterior_mean::AbstractFloat, self::Node, value_child::Node)

Calculates a node's posterior mean for a VAPE coupling with a single value child.

HGF.calculate_posterior_mean_vape — Method

calculate_posterior_mean_vape(posterior_mean::AbstractFloat, self::Node, value_children::Vector{Node})

Calculates a node's posterior mean for a VAPE coupling with multiple value children.

HGF.calculate_posterior_mean_vope — Method

calculate_posterior_mean_vope(posterior_mean::AbstractFloat, self::Node, volatility_child::Bool)

Calculates a node's posterior mean for a VOPE coupling without a volatility child.

HGF.calculate_posterior_mean_vope — Method

calculate_posterior_mean_vope(posterior_mean::AbstractFloat, self::Node, volatility_child::Node)

Calculates a node's posterior mean for a VOPE coupling with a single volatility child.

HGF.calculate_posterior_mean_vope — Method

calculate_posterior_mean_vope(posterior_mean::AbstractFloat, self::Node, volatility_children::Vector{Node})

Calculates a node's posterior mean for a VOPE coupling with multiple volatility children.

HGF.calculate_posterior_precision — Method

calculate_posterior_precision(self::Node, value_children, volatility_children)

Calculates a node's posterior precision.

HGF.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(posterior_precision::AbstractFloat, self::Node, value_child::Bool)

Calculates a node's posterior precision for a VAPE coupling without a value child.

HGF.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(posterior_precision::AbstractFloat, self::Node, value_child::Node)

Calculates a node's posterior precision for a VAPE coupling with a single value child.

HGF.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(posterior_precision::AbstractFloat, self::Node, value_children::Vector{Node})

Calculates a node's posterior precision for a VAPE coupling with multiple value children.

HGF.calculate_posterior_precision_vope — Method

calculate_posterior_precision_vope(posterior_precision::AbstractFloat, self::Node, volatility_child::Bool)

Calculates a node's posterior precision for a VOPE coupling without a volatility child.

HGF.calculate_posterior_precision_vope — Method

calculate_posterior_precision_vope(posterior_precision::AbstractFloat, self::Node, volatility_child::Node)

Calculates a node's posterior precision for a VOPE coupling with a single volatility child.

HGF.calculate_posterior_precision_vope — Method

calculate_posterior_precision_vope(posterior_precision::AbstractFloat, self::Node, volatility_children::Vector{Node})

Calculates a node's posterior precision for a VOPE coupling with multiple volatility children.

HGF.calculate_posterior_precision_vope_helper — Method

calculate_posterior_precision_vope_helper(
    auxiliary_prediction_precision::AbstractFloat,
    child_volatility_coupling::AbstractFloat,
    child_volatility_prediction_error::AbstractFloat)

Helper function which calculates the additive term for updating posterior precision in a VOPE coupling.

HGF.calculate_prediction_mean — Method

calculate_prediction_mean(self::Node, value_parent::Bool)

Calculates a node's prediction mean without a value parent.

HGF.calculate_prediction_mean — Method

calculate_prediction_mean(self::Node, value_parent::Node)

Calculates a node's prediction mean with a single value parent.

HGF.calculate_prediction_mean — Method

calculate_prediction_mean(self::Node, value_parent::Vector{Node})

Calculates a node's prediction mean with multiple value parents.

HGF.calculate_prediction_precision — Method

calculate_prediction_precision(self::Node)

Calculates a node's prediction precision.

HGF.calculate_prediction_volatility — Method

calculate_prediction_volatility(self::Node, volatility_parent::Bool)

Calculates a node's prediction volatility without a volatility parent.

HGF.calculate_prediction_volatility — Method

calculate_prediction_volatility(self::Node, volatility_parent::Node)

Calculates a node's prediction volatility with a single value parent.

HGF.calculate_prediction_volatility — Method

calculate_prediction_volatility(self::Node,  volatility_parents::Vector{Node})

Calculates a node's prediction volatility with multiple volatility parents.

HGF.calculate_value_prediction_error — Method

calculate_value_prediction_error(self::Node)

Calculate's a node's value prediction error.

HGF.calculate_volatility_prediction_error — Method

calculate_volatility_prediction_error(self::Node)

Calculates a node's volatility prediction error.

HGF.update_node — Method

update_node(self::InputNode)

Full update function for an input node.

HGF.update_node — Method

update_node(self::Node)

Full update function for a single node. States, parents and children are contained within the node.