HierarchicalGaussianFiltering

Documentation for HierarchicalGaussianFiltering.

HierarchicalGaussianFiltering

Theory

Using the Package

Tutorials

List of Functions

The generative model of the HGF: Volatility vs. value coupling

Belief updates in the HGF: Computations of nodes

Computations for VAPE coupling

Update Step

Prediction Error Step

Prediction Step

Computations for VOPE coupling

Update Step

Prediction Error Step

Prediction Step

Theory

Generative model

Node

VAPE

VOPE

Using the Package

Tutorials

List of Functions

HierarchicalGaussianFiltering.AbstractInputNode

HierarchicalGaussianFiltering.AbstractNode

HierarchicalGaussianFiltering.AbstractStateNode

HierarchicalGaussianFiltering.BinaryInputNode

HierarchicalGaussianFiltering.BinaryInputNodeHistory

HierarchicalGaussianFiltering.BinaryInputNodeState

HierarchicalGaussianFiltering.BinaryStateNode

HierarchicalGaussianFiltering.BinaryStateNodeHistory

HierarchicalGaussianFiltering.BinaryStateNodeState

HierarchicalGaussianFiltering.CategoricalInputNode

HierarchicalGaussianFiltering.CategoricalInputNodeHistory

HierarchicalGaussianFiltering.CategoricalInputNodeState

HierarchicalGaussianFiltering.CategoricalStateNode

HierarchicalGaussianFiltering.CategoricalStateNodeHistory

HierarchicalGaussianFiltering.CategoricalStateNodeState

HierarchicalGaussianFiltering.ContinuousInputNode

HierarchicalGaussianFiltering.ContinuousInputNodeHistory

HierarchicalGaussianFiltering.ContinuousInputNodeParams

HierarchicalGaussianFiltering.ContinuousInputNodeState

HierarchicalGaussianFiltering.ContinuousStateNode

HierarchicalGaussianFiltering.ContinuousStateNodeHistory

HierarchicalGaussianFiltering.ContinuousStateNodeParams

HierarchicalGaussianFiltering.ContinuousStateNodeState

HierarchicalGaussianFiltering.HGF

HierarchicalGaussianFiltering.OrderedNodes

ActionModels.get_history

ActionModels.get_history

ActionModels.get_history

ActionModels.get_history

ActionModels.get_history

ActionModels.get_states

ActionModels.get_states

ActionModels.get_states

ActionModels.get_states

ActionModels.get_states

ActionModels.give_inputs!

ActionModels.give_inputs!

ActionModels.give_inputs!

ActionModels.plot_predictive_simulation

ActionModels.plot_trajectory

ActionModels.plot_trajectory

ActionModels.plot_trajectory!

ActionModels.plot_trajectory!

ActionModels.premade_agent

ActionModels.reset!

ActionModels.set_params!

HierarchicalGaussianFiltering.calculate_auxiliary_prediction_precision

HierarchicalGaussianFiltering.calculate_auxiliary_prediction_precision

HierarchicalGaussianFiltering.calculate_posterior

HierarchicalGaussianFiltering.calculate_posterior_mean

HierarchicalGaussianFiltering.calculate_posterior_mean

HierarchicalGaussianFiltering.calculate_posterior_mean_value_child_increment

HierarchicalGaussianFiltering.calculate_posterior_mean_value_child_increment

HierarchicalGaussianFiltering.calculate_posterior_mean_volatility_child_increment

HierarchicalGaussianFiltering.calculate_posterior_precision

HierarchicalGaussianFiltering.calculate_posterior_precision

HierarchicalGaussianFiltering.calculate_posterior_precision_vape

HierarchicalGaussianFiltering.calculate_posterior_precision_vape

HierarchicalGaussianFiltering.calculate_posterior_precision_vope

HierarchicalGaussianFiltering.calculate_prediction

HierarchicalGaussianFiltering.calculate_prediction_mean

HierarchicalGaussianFiltering.calculate_prediction_mean

HierarchicalGaussianFiltering.calculate_prediction_precision

HierarchicalGaussianFiltering.calculate_prediction_precision

HierarchicalGaussianFiltering.calculate_prediction_precision

HierarchicalGaussianFiltering.calculate_prediction_volatility

HierarchicalGaussianFiltering.calculate_value_prediction_error

HierarchicalGaussianFiltering.calculate_value_prediction_error

HierarchicalGaussianFiltering.calculate_value_prediction_error

HierarchicalGaussianFiltering.calculate_value_prediction_error

HierarchicalGaussianFiltering.calculate_volatility_prediction_error

HierarchicalGaussianFiltering.calculate_volatility_prediction_error

HierarchicalGaussianFiltering.check_hgf

HierarchicalGaussianFiltering.check_hgf

HierarchicalGaussianFiltering.check_hgf

HierarchicalGaussianFiltering.check_hgf

HierarchicalGaussianFiltering.check_hgf

HierarchicalGaussianFiltering.check_hgf

HierarchicalGaussianFiltering.check_hgf

HierarchicalGaussianFiltering.enter_node_inputs!

HierarchicalGaussianFiltering.enter_node_inputs!

HierarchicalGaussianFiltering.enter_node_inputs!

HierarchicalGaussianFiltering.get_prediction

HierarchicalGaussianFiltering.get_prediction

HierarchicalGaussianFiltering.get_prediction

HierarchicalGaussianFiltering.get_surprise

HierarchicalGaussianFiltering.get_surprise

HierarchicalGaussianFiltering.get_surprise

HierarchicalGaussianFiltering.get_surprise

HierarchicalGaussianFiltering.get_surprise

HierarchicalGaussianFiltering.get_surprise

HierarchicalGaussianFiltering.get_surprise

HierarchicalGaussianFiltering.hgf_binary_softmax_action

HierarchicalGaussianFiltering.hgf_gaussian_action

HierarchicalGaussianFiltering.hgf_predict_category_action

HierarchicalGaussianFiltering.hgf_unit_square_sigmoid_action

HierarchicalGaussianFiltering.init_hgf

HierarchicalGaussianFiltering.init_node

HierarchicalGaussianFiltering.premade_JGET

HierarchicalGaussianFiltering.premade_binary_2level

HierarchicalGaussianFiltering.premade_binary_3level

HierarchicalGaussianFiltering.premade_continuous_2level

HierarchicalGaussianFiltering.premade_hgf

HierarchicalGaussianFiltering.premade_hgf_binary_softmax

HierarchicalGaussianFiltering.premade_hgf_gaussian

HierarchicalGaussianFiltering.premade_hgf_multiple_actions

HierarchicalGaussianFiltering.premade_hgf_predict_category

HierarchicalGaussianFiltering.premade_hgf_unit_square_sigmoid

HierarchicalGaussianFiltering.update_hgf!

HierarchicalGaussianFiltering.update_hgf_binary_softmax_action

HierarchicalGaussianFiltering.update_hgf_gaussian_action

HierarchicalGaussianFiltering.update_hgf_multiple_actions

HierarchicalGaussianFiltering.update_hgf_predict_category_action

HierarchicalGaussianFiltering.update_hgf_unit_square_sigmoid_action

HierarchicalGaussianFiltering.update_node_input!

HierarchicalGaussianFiltering.update_node_posterior!

HierarchicalGaussianFiltering.update_node_posterior!

HierarchicalGaussianFiltering.update_node_prediction!

HierarchicalGaussianFiltering.update_node_prediction!

HierarchicalGaussianFiltering.update_node_prediction!

HierarchicalGaussianFiltering.update_node_prediction!

HierarchicalGaussianFiltering.update_node_prediction!

HierarchicalGaussianFiltering.update_node_prediction!

HierarchicalGaussianFiltering.update_node_value_prediction_error!

HierarchicalGaussianFiltering.update_node_value_prediction_error!

HierarchicalGaussianFiltering.update_node_value_prediction_error!

HierarchicalGaussianFiltering.update_node_value_prediction_error!

HierarchicalGaussianFiltering.update_node_volatility_prediction_error!

HierarchicalGaussianFiltering.update_node_volatility_prediction_error!

HierarchicalGaussianFiltering.update_node_volatility_prediction_error!

HierarchicalGaussianFiltering.update_node_volatility_prediction_error!

HierarchicalGaussianFiltering.update_node_volatility_prediction_error!

HierarchicalGaussianFiltering.update_node_volatility_prediction_error!

MCMCChains.get_params

MCMCChains.get_params

MCMCChains.get_params

MCMCChains.get_params

MCMCChains.get_params

MCMCChains.get_params

HierarchicalGaussianFiltering.AbstractInputNode — Type

HierarchicalGaussianFiltering.AbstractNode — Type

HierarchicalGaussianFiltering.AbstractStateNode — Type

HierarchicalGaussianFiltering.BinaryInputNode — Type

HierarchicalGaussianFiltering.BinaryInputNodeHistory — Type

HierarchicalGaussianFiltering.BinaryInputNodeState — Type

HierarchicalGaussianFiltering.BinaryStateNode — Type

HierarchicalGaussianFiltering.BinaryStateNodeHistory — Type

HierarchicalGaussianFiltering.BinaryStateNodeState — Type

HierarchicalGaussianFiltering.CategoricalInputNode — Type

HierarchicalGaussianFiltering.CategoricalInputNodeHistory — Type

HierarchicalGaussianFiltering.CategoricalInputNodeState — Type

HierarchicalGaussianFiltering.CategoricalStateNode — Type

HierarchicalGaussianFiltering.CategoricalStateNodeHistory — Type

HierarchicalGaussianFiltering.CategoricalStateNodeState — Type

HierarchicalGaussianFiltering.ContinuousInputNode — Type

HierarchicalGaussianFiltering.ContinuousInputNodeHistory — Type

HierarchicalGaussianFiltering.ContinuousInputNodeParams — Type

HierarchicalGaussianFiltering.ContinuousInputNodeState — Type

HierarchicalGaussianFiltering.ContinuousStateNode — Type

HierarchicalGaussianFiltering.ContinuousStateNodeHistory — Type

HierarchicalGaussianFiltering.ContinuousStateNodeParams — Type

HierarchicalGaussianFiltering.ContinuousStateNodeState — Type

HierarchicalGaussianFiltering.HGF — Type

HierarchicalGaussianFiltering.OrderedNodes — Type

ActionModels.get_history — Method

ActionModels.get_history — Method

ActionModels.get_history — Method

ActionModels.get_history — Method

ActionModels.get_history — Method

ActionModels.get_states — Method

ActionModels.get_states — Method

ActionModels.get_states — Method

ActionModels.get_states — Method

ActionModels.get_states — Method

ActionModels.give_inputs! — Method

give_inputs!(hgf::HGF, inputs::Array)

Function for inputting multiple observations to an hgf. Input is structured as an array, with one column per input node and one row per timestep.

ActionModels.give_inputs! — Method

give_inputs!(hgf::HGF, inputs::Number)

Function for inputting multiple observations to an hgf. Input is a single value.

ActionModels.give_inputs! — Method

ActionModels.plot_predictive_simulation — Method

ActionModels.plot_trajectory! — Method

ActionModels.plot_trajectory! — Method

ActionModels.plot_trajectory — Method

ActionModels.plot_trajectory — Method

ActionModels.premade_agent — Function

premade_agent(model_name::String, hgf::HGF, params_list::NamedTuple = (;))

Function for making a premade agent, where a HGF is passed as a separate argument.

ActionModels.reset! — Method

ActionModels.set_params! — Method

HierarchicalGaussianFiltering.calculate_auxiliary_prediction_precision — Method

calculate_auxiliary_prediction_precision(node::AbstractInputNode)

An input nodes auxiliary precision is always 1.

HierarchicalGaussianFiltering.calculate_auxiliary_prediction_precision — Method

calculate_auxiliary_prediction_precision(node::AbstractNode)

Calculates a node's auxiliary prediction precision.

HierarchicalGaussianFiltering.calculate_posterior — Method

HierarchicalGaussianFiltering.calculate_posterior_mean — Method

calculate_posterior_mean(node::AbstractNode, value_children, volatility_children)

Calculates a node's posterior mean.

HierarchicalGaussianFiltering.calculate_posterior_mean — Method

calculate_posterior_mean(node::BinaryStateNode, value_children, volatility_children)

Calculates a node's posterior mean.

HierarchicalGaussianFiltering.calculate_posterior_mean_value_child_increment — Method

calculate_posterior_mean_value_child_increment(
    node::AbstractNode,
    child::AbstractNode)

Calculates the posterior mean update term for a single continuous value child to a state node.

HierarchicalGaussianFiltering.calculate_posterior_mean_value_child_increment — Method

calculate_posterior_mean_value_child_increment(
    node::AbstractNode,
    child::BinaryStateNode)

Calculates the posterior mean update term for a single binary value child to a state node.

HierarchicalGaussianFiltering.calculate_posterior_mean_volatility_child_increment — Method

calculate_posterior_mean_volatility_child_increment(
    node::AbstractNode,
    child::Any)

Calculates the posterior mean update term for a single continuos volatility child to a state node.

HierarchicalGaussianFiltering.calculate_posterior_precision — Method

calculate_posterior_precision(
    node::AbstractNode,
    value_children,
    volatility_children)

Calculates a node's posterior precision.

HierarchicalGaussianFiltering.calculate_posterior_precision — Method

calculate_posterior_precision(
    node::BinaryStateNode,
    value_children,
    volatility_children)

Calculates a binary node's posterior precision.

HierarchicalGaussianFiltering.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(
    node::AbstractNode,
    child::AbstractNode)

Calculates the posterior precision update term for a single continuous value child to a state node.

HierarchicalGaussianFiltering.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(
    node::AbstractNode,
    child::BinaryStateNode)

Calculates the posterior precision update term for a single binary value child to a state node.

HierarchicalGaussianFiltering.calculate_posterior_precision_vope — Method

calculate_posterior_precision_vope(
    posterior_precision::Real,
    node::AbstractNode,
    volatility_children::Any)

Calculates the posterior precision update term for a single continuous volatility child to a state node.

HierarchicalGaussianFiltering.calculate_prediction — Method

HierarchicalGaussianFiltering.calculate_prediction_mean — Method

calculate_prediction_mean(node::AbstractNode, value_parent::Any)

Calculates a node's prediction mean.

HierarchicalGaussianFiltering.calculate_prediction_mean — Method

calculate_prediction_mean(node::AbstractNode, value_parent::Any)

Calculates a binary state node's prediction mean.

HierarchicalGaussianFiltering.calculate_prediction_precision — Method

calculate_prediction_precision(node::ContinuousInputNode)

Calculates an input node's prediction precision.

HierarchicalGaussianFiltering.calculate_prediction_precision — Method

calculate_prediction_precision(node::ContinuousStateNode)

Calculates a node's prediction precision.

HierarchicalGaussianFiltering.calculate_prediction_precision — Method

calculate_prediction_precision(node::BinaryStateNode)

Calculates a binary state node's prediction precision.

HierarchicalGaussianFiltering.calculate_prediction_volatility — Method

calculate_prediction_volatility(node::AbstractNode,  volatility_parents::Any)

Calculates a node's prediction volatility.

HierarchicalGaussianFiltering.calculate_value_prediction_error — Method

calculate_value_prediction_error(node::AbstractNode)

Calculate's a state node's value prediction error.

HierarchicalGaussianFiltering.calculate_value_prediction_error — Method

calculate_value_prediction_error(node::BinaryInputNode)

Calculates the prediciton error of a binary input node with finite precision

HierarchicalGaussianFiltering.calculate_value_prediction_error — Method

HierarchicalGaussianFiltering.calculate_value_prediction_error — Method

calculate_value_prediction_error(node::ContinuousInputNode, value_parents::Any)

Calculate's an input node's value prediction error.

HierarchicalGaussianFiltering.calculate_volatility_prediction_error — Method

calculate_volatility_prediction_error(node::AbstractNode)

Calculates a state node's volatility prediction error.

HierarchicalGaussianFiltering.calculate_volatility_prediction_error — Method

calculate_volatility_prediction_error(node::ContinuousInputNode, value_parents::Any)

Calculates an input node's volatility prediction error.

HierarchicalGaussianFiltering.check_hgf — Method

check_hgf(node::BinaryInputNode)

Function for checking the validity of a single binary input node

HierarchicalGaussianFiltering.check_hgf — Method

check_hgf(node::BinaryStateNode)

Function for checking the validity of a single binary state node

HierarchicalGaussianFiltering.check_hgf — Method

check_hgf(node::CategoricalInputNode)

Function for checking the validity of a single binary input node

HierarchicalGaussianFiltering.check_hgf — Method

check_hgf(node::CategoricalStateNode)

Function for checking the validity of a single categorical state node

HierarchicalGaussianFiltering.check_hgf — Method

check_hgf(node::ContinuousInputNode)

Function for checking the validity of a single continuous input node

HierarchicalGaussianFiltering.check_hgf — Method

check_hgf(node::ContinuousStateNode)

Function for checking the validity of a single continous state node

HierarchicalGaussianFiltering.check_hgf — Method

Function for checking if the specified HGF structure is valid

HierarchicalGaussianFiltering.enter_node_inputs! — Method

enter_node_inputs!(hgf::HGF, inputs::Dict)

Function for entering multiple inputs, structured as a dictionary, to multiple input nodes.

HierarchicalGaussianFiltering.enter_node_inputs! — Method

enter_node_inputs!(hgf::HGF, input::Number)

Function for entering a single input to a single input node. Can either take a single number, or a tuple which also includes the precision of the input.

HierarchicalGaussianFiltering.enter_node_inputs! — Method

enter_node_inputs!(hgf::HGF, inputs::Vector)

Function for entering multiple inputs, structured as a vector, to multiple input nodes.

HierarchicalGaussianFiltering.get_prediction — Function

HierarchicalGaussianFiltering.get_prediction — Function

get_prediction(hgf::HGF, node_name::String)

Gets the full prediction for the next timestep for a specified node in an HGF.

HierarchicalGaussianFiltering.get_prediction — Method

get_prediction(node::AbstractNode)

Gets the full prediction for the next timestep for a single node.

HierarchicalGaussianFiltering.get_surprise — Method

HierarchicalGaussianFiltering.get_surprise — Method

HierarchicalGaussianFiltering.get_surprise — Method

get_surprise(node::BinaryInputNode)

Calculates the surprise of a binary input node on seeing the last input.

HierarchicalGaussianFiltering.get_surprise — Method

get_surprise(node::CategoricalInputNode)

Calculates the surprise of a categorical input node on seeing the last input.

HierarchicalGaussianFiltering.get_surprise — Method

get_surprise(node::ContinuousInputNode)

Calculates the surprise of an input node on seeing the last input. Implements the equation: −log(p(u(k)))= 1(log(2π)−log(πˆ(k))+πˆ(k)(u(k) −μˆ(k) )2)

HierarchicalGaussianFiltering.get_surprise — Method

getsurprise(hgf::HGF, nodename::String = "u")

Calculates the surprisal of a specified input node in an HGF.

HierarchicalGaussianFiltering.get_surprise — Method

getsurprise(hgf::HGF, nodename::String = "u")

Calculates the surprisal of a specified input node in an HGF.

HierarchicalGaussianFiltering.hgf_binary_softmax_action — Method

hgf_binary_softmax_action(agent, input)

Action model which gives a binary action. The action probability is the softmax of a specified state of a node.

HierarchicalGaussianFiltering.hgf_gaussian_action — Method

hgf_gaussian_action(agent, input)

Action model which reports a given HGF state with Gaussian noise.

HierarchicalGaussianFiltering.hgf_predict_category_action — Method

unit_square_sigmoid_action(agent, input)

Action model which gives a binary action. The action probability is the unit square sigmoid of a specified state of a node.

HierarchicalGaussianFiltering.hgf_unit_square_sigmoid_action — Method

unit_square_sigmoid_action(agent, input)

Action model which gives a binary action. The action probability is the unit square sigmoid of a specified state of a node.

HierarchicalGaussianFiltering.init_hgf — Method

function init_hgf(
    default_params,
    input_nodes,
    state_nodes,
    edges,
    update_order = false,
)

Function for initializing the structure of an HGF model.

HierarchicalGaussianFiltering.init_node — Method

init_node(input_or_state_node, node_defaults, node_info)

Function for creating a node, given specifications

HierarchicalGaussianFiltering.premade_JGET — Method

premade_JGET(params_list, starting_state_list)

The JGET model. It has a single continuous input node u, with a value parent x1, and a volatility parent x3. x1 has volatility parent x2, and x3 has a volatility parent x4.

HierarchicalGaussianFiltering.premade_binary_2level — Method

premade_binary_2level(params_list, starting_state_list)

The standard binary 2 level HGF model

HierarchicalGaussianFiltering.premade_binary_3level — Method

premade_binary_3level(params_list, starting_state_list)

The standard binary 3 level HGF model

HierarchicalGaussianFiltering.premade_continuous_2level — Method

premade_continuous_2level(params_list, starting_state_list)

The standard 2 level HGF. It has a continous input node U, with a single value parent x1, which in turn has a single volatility parent x2.

HierarchicalGaussianFiltering.premade_hgf — Function

function premade_hgf(
    model_name,
    params,
    starting_state,
)

Function for initializing the structure of an HGF model.

HierarchicalGaussianFiltering.premade_hgf_binary_softmax — Method

premade_hgf_binary_softmax(
    hgf = premade_hgf("binary_3level"),
    action_precision = 1,
    target_node = "x1",
    target_state = "posterior_mean",
)

Function that initializes as premade HGF binary softmax action agent

HierarchicalGaussianFiltering.premade_hgf_gaussian — Method

premade_hgf_gaussian(
    hgf = premade_hgf("continuous_2level"),
    action_precision = 1,
    target_node = "x1",
    target_state = "posterior_mean",
)

Function that initializes as premade HGF gaussian action agent

HierarchicalGaussianFiltering.premade_hgf_multiple_actions — Method

HierarchicalGaussianFiltering.premade_hgf_predict_category — Method

HierarchicalGaussianFiltering.premade_hgf_unit_square_sigmoid — Method

premade_hgf_unit_square_sigmoid(
    hgf = premade_hgf("binary_3level"),
    action_precision = 1,
    target_node = "x1",
    target_state = "posterior_mean",
)

Function that initializes as premade HGF binary softmax action agent

HierarchicalGaussianFiltering.update_hgf! — Method

update_hgf!(hgf::HGF, inputs)

Function for updating all nodes in an HGF hierarchy.

HierarchicalGaussianFiltering.update_hgf_binary_softmax_action — Method

HierarchicalGaussianFiltering.update_hgf_gaussian_action — Method

HierarchicalGaussianFiltering.update_hgf_multiple_actions — Method

HierarchicalGaussianFiltering.update_hgf_predict_category_action — Method

HierarchicalGaussianFiltering.update_hgf_unit_square_sigmoid_action — Method

HierarchicalGaussianFiltering.update_node_input! — Method

update_node_input!(node::AbstractInputNode, input::Real)

Function for updating the input for a single input node

HierarchicalGaussianFiltering.update_node_posterior! — Method

update_node_posterior!(node::ContinuousStateNode)

Function for updating the posterior of a single state node

HierarchicalGaussianFiltering.update_node_posterior! — Method

update_node_posterior!(node::CategoricalStateNode)

Function for updating the posterior of a single categorical state node

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::ContinuousInputNode)

Function for updating the posterior of a single input node

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::ContinuousStateNode)

Function for updating the prediction for a single state node

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::BinaryInputNode)

The prediction precision is constant for binary input nodes, so nothing is done here.

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::BinaryStateNode)

Function for updating the prediction for a single Binary state node

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::BinaryInputNode)

The prediction precision is constant for categorical input nodes, so nothing is done here.

HierarchicalGaussianFiltering.update_node_prediction! — Method

update_node_prediction!(node::CategoricalStateNode)

Function for updating the prediction for a single Cateogrical state node

HierarchicalGaussianFiltering.update_node_value_prediction_error! — Method

update_node_value_prediction_error!(node::ContinuousInputNode)

Function for updating the value prediction error of a single input node

HierarchicalGaussianFiltering.update_node_value_prediction_error! — Method

Function for updating the value prediction error of a single state node

HierarchicalGaussianFiltering.update_node_value_prediction_error! — Method

update_node_value_prediction_error!(node::BinaryInputNode)

Function for updating the value prediction error of a single input node.

HierarchicalGaussianFiltering.update_node_value_prediction_error! — Method

update_node_value_prediction_error!(node::CategoricalInputNode)

Function for updating the value prediction error of a single input node.

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(node::ContinuousInputNode)

Function for updating the volatility prediction error of a single input node

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

Function for updating the volatility prediction error of a single state node

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(node::BinaryInputNode)

There is no volatility prediction error for binary input nodes.

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

Function for updating the volatility prediction error of a single binary state node

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

update_node_volatility_prediction_error!(node::CategoricalInputNode)

There is no volatility prediction error for binary input nodes.

HierarchicalGaussianFiltering.update_node_volatility_prediction_error! — Method

Function for updating the volatility prediction error of a single cateogrical state node

MCMCChains.get_params — Method

MCMCChains.get_params — Method

MCMCChains.get_params — Method

MCMCChains.get_params — Method

MCMCChains.get_params — Method

MCMCChains.get_params — Method