HGF

Documentation for HGF.

HGF.calculate_auxiliary_prediction_precision

HGF.calculate_posterior_mean

HGF.calculate_posterior_mean_vape

HGF.calculate_posterior_mean_vope

HGF.calculate_posterior_precision

HGF.calculate_posterior_precision_vape

HGF.calculate_posterior_precision_vope

HGF.calculate_posterior_precision_vope_helper

HGF.calculate_prediction_mean

HGF.calculate_prediction_precision

HGF.calculate_prediction_precision

HGF.calculate_prediction_volatility

HGF.calculate_value_prediction_error

HGF.calculate_value_prediction_error

HGF.calculate_volatility_prediction_error

HGF.calculate_volatility_prediction_error

HGF.init_HGF

HGF.update_hierarchy

HGF.update_hierarchy

HGF.update_node

HGF.update_node

HGF.calculate_auxiliary_prediction_precision — Method

calculate_auxiliary_prediction_precision(self::AbstractNode)

Calculates a node's auxiliary prediction precision.

HGF.calculate_posterior_mean — Method

calculate_posterior_mean(self::AbstractNode, value_children, volatility_children)

Calculates a node's posterior mean.

HGF.calculate_posterior_mean_vape — Method

calculate_posterior_mean_vape(
    posterior_mean::AbstractFloat,
    self::AbstractNode,
    value_children::Any)

Calculates a node's posterior mean for a VAPE coupling.

HGF.calculate_posterior_mean_vope — Method

calculate_posterior_mean_vope(
    posterior_mean::AbstractFloat,
    self::AbstractNode,
    volatility_children::Any)

Calculates a node's posterior mean for a VOPE coupling.

HGF.calculate_posterior_precision — Method

calculate_posterior_precision(
    self::AbstractNode,
    value_children,
    volatility_children)

Calculates a node's posterior precision.

HGF.calculate_posterior_precision_vape — Method

calculate_posterior_precision_vape(
    posterior_precision::AbstractFloat,
    self::AbstractNode,
    value_children::Any)

Calculates a node's posterior precision for a VAPE coupling.

HGF.calculate_posterior_precision_vope — Method

calculate_posterior_precision_vope(
    posterior_precision::AbstractFloat,
    self::AbstractNode,
    volatility_children::Any)

Calculates a node's posterior precision for a VOPE coupling.

HGF.calculate_posterior_precision_vope_helper — Method

calculate_posterior_precision_vope_helper(
    auxiliary_prediction_precision::AbstractFloat,
    child_volatility_coupling::AbstractFloat,
    child_volatility_prediction_error::AbstractFloat)

Helper function which calculates the additive term for updating posterior precision in a VOPE coupling.

HGF.calculate_prediction_mean — Method

calculate_prediction_mean(self::AbstractNode, value_parent::Vector{Any})

Calculates a node's prediction mean.

HGF.calculate_prediction_precision — Method

calculate_prediction_precision(self::InputNode)

Calculates an input node's prediction precision.

HGF.calculate_prediction_precision — Method

calculate_prediction_precision(self::StateNode)

Calculates a node's prediction precision.

HGF.calculate_prediction_volatility — Method

calculate_prediction_volatility(self::AbstractNode,  volatility_parents::Vector{Any})

Calculates a node's prediction volatility.

HGF.calculate_value_prediction_error — Method

calculate_value_prediction_error(self::AbstractNode)

Calculate's a state node's value prediction error.

HGF.calculate_value_prediction_error — Method

calculate_value_prediction_error(self::InputNode, value_parents::Vector{Any})

Calculate's an input node's value prediction error.

HGF.calculate_volatility_prediction_error — Method

calculate_volatility_prediction_error(self::AbstractNode)

Calculates a state node's volatility prediction error.

HGF.calculate_volatility_prediction_error — Method

calculate_volatility_prediction_error(self::InputNode, value_parents::Vector{Any})

Calculates an input node's volatility prediction error.

HGF.init_HGF — Function

function init_HGF(
    default_params,
    input_nodes,
    state_nodes,
    child_parent_relations,
    update_order = false,
)

Function for initializing the structure of an HGF model.

HGF.update_hierarchy — Method

update_hierarchy(HGF_struct::HGFModel, input::AbstractFloat)

Function for updating all nodes in an HGF hierarchy, with a single input node.

HGF.update_hierarchy — Method

update_hierarchy(HGF_struct::HGFModel, input::Dict{String, Float64})

Function for updating all nodes in an HGF hierarchy, with multiple input nodes.

HGF.update_node — Method

update_node(self::InputNode, input::AbstractFloat)

Full update function for a continuous input node.

HGF.update_node — Method

update_node(self::StateNode)

Full update function for a single node. States, parents and children are contained within the node.